+++ crypto/keys/secp256k1/secp256k1_cgo.go
@@ -4,14 +4,13 @@
 package secp256k1
 
 import (
-	"github.com/cometbft/cometbft/crypto"
-
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1/internal/secp256k1"
+	"github.com/cosmos/crypto/hash/sha256"
 )
 
 // Sign creates an ECDSA signature on curve Secp256k1, using SHA256 on the msg.
 func (privKey *PrivKey) Sign(msg []byte) ([]byte, error) {
-	rsv, err := secp256k1.Sign(crypto.Sha256(msg), privKey.Key)
+	rsv, err := secp256k1.Sign(sha256.Sum(msg), privKey.Key)
 	if err != nil {
 		return nil, err
 	}
@@ -23,5 +22,5 @@ func (privKey *PrivKey) Sign(msg []byte) ([]byte, error) {
 // VerifySignature validates the signature.
 // The msg will be hashed prior to signature verification.
 func (pubKey *PubKey) VerifySignature(msg, sigStr []byte) bool {
-	return secp256k1.VerifySignature(pubKey.Bytes(), crypto.Sha256(msg), sigStr)
+	return secp256k1.VerifySignature(pubKey.Bytes(), sha256.Sum(msg), sigStr)
 }